+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.9.3
+ - Fixed a bug where onAttached was called before the node was fully attached.
  - Made the remove method cascade, so now all descendants of the node will also be fed to the onRemoved callback.
  - Made the remove method also call the moved callback for potentially displaced siblings
 ## 0.9.2